+++ test/dummy/rails50_sprockets3/app/controllers/admin_controller.rb
@@ -3,13 +3,11 @@ class AdminController < Super::ApplicationController
 
   def site_navigation
     Super::Navigation.new do |nav|
-      [
-        nav.link(Member),
-        nav.link(Ship),
-        nav.menu("Other")[
-          nav.rest
-        ]
-      ]
+      nav.link(Member)
+      nav.link(Ship)
+      nav.menu("Other") do
+        nav.rest
+      end
     end
   end
 end